move base type tests
