+++ api-js/src/types/base/__tests__/dkim-connection-query.test.js
@@ -4,18 +4,18 @@ const { ArangoTools, dbNameFromFile } = require('arango-tools')
 const { toGlobalId } = require('graphql-relay')
 const { graphql, GraphQLSchema } = require('graphql')
 
-const { createQuerySchema } = require('../queries')
-const { createMutationSchema } = require('../mutations')
-const { makeMigrations } = require('../../migrations')
-const { cleanseInput } = require('../validators')
-const { checkDomainPermission, userRequired } = require('../auth')
+const { createQuerySchema } = require('../../../queries')
+const { createMutationSchema } = require('../../../mutations')
+const { makeMigrations } = require('../../../../migrations')
+const { cleanseInput } = require('../../../validators')
+const { checkDomainPermission, userRequired } = require('../../../auth')
 const {
   dkimLoaderConnectionsByDomainId,
   dkimLoaderByKey,
   domainLoaderByDomain,
   domainLoaderByKey,
   userLoaderByKey,
-} = require('../loaders')
+} = require('../../../loaders')
 
 describe('given the dkimType object', () => {
   let query,
