+++ api-js/src/types/base/__tests__/https-connection-query.test.js
@@ -4,22 +4,22 @@ const { ArangoTools, dbNameFromFile } = require('arango-tools')
 const { toGlobalId } = require('graphql-relay')
 const { graphql, GraphQLSchema } = require('graphql')
 
-const { createQuerySchema } = require('../queries')
-const { createMutationSchema } = require('../mutations')
-const { makeMigrations } = require('../../migrations')
-const { cleanseInput } = require('../validators')
-const { checkDomainPermission, userRequired } = require('../auth')
+const { createQuerySchema } = require('../../../queries')
+const { createMutationSchema } = require('../../../mutations')
+const { makeMigrations } = require('../../../../migrations')
+const { cleanseInput } = require('../../../validators')
+const { checkDomainPermission, userRequired } = require('../../../auth')
 const {
-  dmarcLoaderConnectionsByDomainId,
-  dmarcLoaderByKey,
-  dmarcGuidanceTagLoader,
-  dmarcGuidanceTagConnectionsLoader,
+  httpsLoaderByKey,
+  httpsLoaderConnectionsByDomainId,
+  httpsGuidanceTagLoader,
+  httpsGuidanceTagConnectionsLoader,
   domainLoaderByDomain,
   domainLoaderByKey,
   userLoaderByKey,
-} = require('../loaders')
+} = require('../../../loaders')
 
-describe('given the dmarcType object', () => {
+describe('given the https gql object', () => {
   let query,
     drop,
     truncate,
@@ -29,7 +29,7 @@ describe('given the dmarcType object', () => {
     domain,
     schema,
     org,
-    dmarc
+    https
 
   const consoleInfoOutput = []
   const mockedInfo = (output) => consoleInfoOutput.push(output)
@@ -56,6 +56,7 @@ describe('given the dmarcType object', () => {
       query: createQuerySchema(),
       mutation: createMutationSchema(),
     })
+
     consoleWarnOutput.length = 0
     consoleErrorOutput.length = 0
     consoleInfoOutput.length = 0
@@ -67,6 +68,7 @@ describe('given the dmarcType object', () => {
       tfaValidated: false,
       emailValidated: false,
     })
+
     org = await collections.organizations.save({
       orgDetails: {
         en: {
@@ -104,22 +106,22 @@ describe('given the dmarcType object', () => {
       _from: org._id,
       _to: domain._id,
     })
-    dmarc = await collections.dmarc.save({
+    https = await collections.https.save({
       timestamp: '2020-10-02T12:43:39Z',
-      dmarcPhase: 1,
-      record: 'txtRecord',
-      pPolicy: 'pPolicy',
-      spPolicy: 'spPolicy',
-      pct: 100,
-      guidanceTags: ['dmarc1'],
+      implementation: 'Valid HTTPS',
+      enforced: 'Strict',
+      hsts: 'HSTS Max Age Too Short',
+      hstsAge: '31622400',
+      preloaded: 'HSTS Preloaded',
+      guidanceTags: ['https1'],
     })
-    await collections.domainsDMARC.save({
+    await collections.domainsHTTPS.save({
       _from: domain._id,
-      _to: dmarc._id,
+      _to: https._id,
     })
-    await collections.dmarcGuidanceTags.save({
-      _key: 'dmarc1',
-      tagName: 'DMARC-TAG',
+    await collections.httpsGuidanceTags.save({
+      _key: 'https1',
+      tagName: 'HTTPS-TAG',
       guidance: 'Some Interesting Guidance',
       refLinksGuide: [
         {
@@ -142,6 +144,17 @@ describe('given the dmarcType object', () => {
 
   describe('all fields can be queried', () => {
     it('resolves all fields', async () => {
+      const { setupI18n } = require('@lingui/core')
+      const englishMessages = require('../../../locale/en/messages')
+      const i18n = setupI18n({
+        language: 'en',
+        locales: ['en'],
+        missing: 'Traduction manquante',
+        catalogs: {
+          en: englishMessages,
+        },
+      })
+
       const response = await graphql(
         schema,
         `
@@ -149,8 +162,8 @@ describe('given the dmarcType object', () => {
             findDomainByDomain(domain: "test.domain.gc.ca") {
               id
               domain
-              email {
-                dmarc(first: 5) {
+              web {
+                https(first: 5) {
                   edges {
                     node {
                       id
@@ -158,11 +171,11 @@ describe('given the dmarcType object', () => {
                         id
                       }
                       timestamp
-                      dmarcPhase
-                      record
-                      pPolicy
-                      spPolicy
-                      pct
+                      implementation
+                      enforced
+                      hsts
+                      hstsAge
+                      preloaded
                       guidanceTags(first: 5) {
                         edges {
                           node {
@@ -220,17 +233,18 @@ describe('given the dmarcType object', () => {
             cleanseInput,
           },
           loaders: {
-            dmarcLoaderConnectionsByDomainId: dmarcLoaderConnectionsByDomainId(
+            httpsLoaderConnectionsByDomainId: httpsLoaderConnectionsByDomainId(
               query,
               user._key,
               cleanseInput,
             ),
-            dmarcLoaderByKey: dmarcLoaderByKey(query),
-            dmarcGuidanceTagLoader: dmarcGuidanceTagLoader(query),
-            dmarcGuidanceTagConnectionsLoader: dmarcGuidanceTagConnectionsLoader(
+            httpsLoaderByKey: httpsLoaderByKey(query),
+            httpsGuidanceTagLoader: httpsGuidanceTagLoader(query),
+            httpsGuidanceTagConnectionsLoader: httpsGuidanceTagConnectionsLoader(
               query,
               user._key,
               cleanseInput,
+              i18n,
             ),
             domainLoaderByDomain: domainLoaderByDomain(query),
             domainLoaderByKey: domainLoaderByKey(query),
@@ -244,28 +258,28 @@ describe('given the dmarcType object', () => {
           findDomainByDomain: {
             id: toGlobalId('domains', domain._key),
             domain: 'test.domain.gc.ca',
-            email: {
-              dmarc: {
+            web: {
+              https: {
                 edges: [
                   {
                     node: {
-                      id: toGlobalId('dmarc', dmarc._key),
+                      id: toGlobalId('https', https._key),
                       domain: {
                         id: toGlobalId('domains', domain._key),
                       },
                       timestamp: new Date('2020-10-02T12:43:39Z'),
-                      dmarcPhase: 1,
-                      record: 'txtRecord',
-                      pPolicy: 'pPolicy',
-                      spPolicy: 'spPolicy',
-                      pct: 100,
+                      implementation: 'Valid HTTPS',
+                      enforced: 'Strict',
+                      hsts: 'HSTS Max Age Too Short',
+                      hstsAge: '31622400',
+                      preloaded: 'HSTS Preloaded',
                       guidanceTags: {
                         edges: [
                           {
                             node: {
-                              id: toGlobalId('guidanceTags', 'dmarc1'),
-                              tagId: 'dmarc1',
-                              tagName: 'DMARC-TAG',
+                              id: toGlobalId('guidanceTags', 'https1'),
+                              tagId: 'https1',
+                              tagName: 'HTTPS-TAG',
                               guidance: 'Some Interesting Guidance',
                               refLinks: [
                                 {
@@ -286,8 +300,8 @@ describe('given the dmarcType object', () => {
                         pageInfo: {
                           hasNextPage: false,
                           hasPreviousPage: false,
-                          startCursor: toGlobalId('guidanceTags', 'dmarc1'),
-                          endCursor: toGlobalId('guidanceTags', 'dmarc1'),
+                          startCursor: toGlobalId('guidanceTags', 'https1'),
+                          endCursor: toGlobalId('guidanceTags', 'https1'),
                         },
                       },
                     },
@@ -297,14 +311,15 @@ describe('given the dmarcType object', () => {
                 pageInfo: {
                   hasNextPage: false,
                   hasPreviousPage: false,
-                  startCursor: toGlobalId('dmarc', dmarc._key),
-                  endCursor: toGlobalId('dmarc', dmarc._key),
+                  startCursor: toGlobalId('https', https._key),
+                  endCursor: toGlobalId('https', https._key),
                 },
               },
             },
           },
         },
       }
+
       expect(response).toEqual(expectedResponse)
       expect(consoleInfoOutput).toEqual([
         `User ${user._key} successfully retrieved domain ${domain._key}.`,
